+++ src/client/src/api-lib/index.js
@@ -1,13 +1,10 @@
 import { getHeight } from "./actions";
 import { deleteBoreholes, patchBoreholes } from "./actions/borehole";
-import { loadSettings, patchSettings } from "./actions/settings";
 import { setAuthentication, unsetAuthentication } from "./actions/user";
 import store, { injectReducer } from "./reducers";
 
 export {
   getHeight,
-  loadSettings,
-  patchSettings,
   setAuthentication,
   unsetAuthentication,
   deleteBoreholes,