+++ src/client/src/pages/settings/layerInterface.ts
@@ -1,8 +1,0 @@
-export interface Layer {
-  queryable: boolean;
-  CRS: string[];
-  Title: string;
-  Abstract: string;
-  Name?: string;
-  Identifier?: string;
-}